+++ eras/babbage/impl/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.14.0.0
 
+* Rename `updateUTxOStateByTxValidity` to `updateUTxOState`
 * Rename rule types and deprecate the old names:
   - `BabbageLEDGER` -> `LEDGER`
   - `BabbageUTXO` -> `UTXO`